Make using Error Handler as PSR-7 middleware optional Allow handling uncaught Exceptions (WIP)
