+++ README.md
@@ -5,7 +5,59 @@ Jasny Error Handler
 [![Scrutinizer Code Quality](https://scrutinizer-ci.com/g/jasny/error-handler/badges/quality-score.png?b=master)](https://scrutinizer-ci.com/g/jasny/error-handler/?branch=master)
 [![Code Coverage](https://scrutinizer-ci.com/g/jasny/error-handler/badges/coverage.png?b=master)](https://scrutinizer-ci.com/g/jasny/error-handler/?branch=master)
 
-Error handler as PSR-7 compatible middleware.
+Error handler with PSR-7 support.
+
+Installation
+---
+
+The Jasny Error Handler package is available on [packagist](https://packagist.org/packages/jasny/error-handler).
+Install it using composer:
+
+    composer require jasny/error-handler
+
+
+Usage
+---
+
+```php
+$errorHandler = new Jasny\ErrorHandler();
+```
+
+Just creating an error handler will do nothing. You can use it for logging, handling fatal errors and as PSR-7 compatible
+middleware.
+
+## Logging
+
+By default the error handler with only catch [Throwables](http://php.net/manual/en/class.throwable.php) and not set the
+[php error handler](http://php.net/set_error_handler).
+
+To log these errors, set the logger using the `setLogger()` method. You can log with any
+[PSR-3 compatible logger](http://www.php-fig.org/psr/psr-3/) like [Monolog](https://github.com/Seldaek/monolog).
+
+The `logUncaught()` method will set the error handler, so warnings and notices can be logged. It may also [register a
+shutdown function](http://php.net/manual/en/function.register-shutdown-function.php) to handle uncatchable fatal
+errors.
+
+```php
+use Monolog\Logger;
+use Monolog\Handler\StreamHandler;
+
+$errorHandler = new Jasny\ErrorHandler();
+
+$log = new Logger('test');
+$log->pushHandler(new StreamHandler('path/to/your.log'));
+
+// Log fatal errors, warnings and uncaught exceptions
+$errorHandler->setLogger($log);
+
+$errorHandler->logUncaught(E_PARSE | E_ERROR | E_WARNING | E_USER_WARNING);
+$errorHandler->logUncaught(Exception::class);
+$errorHandler->logUncaught(Error::class); // PHP7 only
+```
+
+## PSR-7 compatible middleware
+
+The error handler can be used as PSR-7 compatible middleware.
 
 The error will catch [Exceptions](http://php.net/manual/en/class.exception.php) and
 [Errors](http://php.net/manual/en/class.error.php).
@@ -17,42 +69,6 @@ You can use this middleware with:
 * [Expressive](http://framework.zend.com/expressive)
 * [Slim 3](http://www.slimframework.com)
 
-You can log with a [PSR-3 compatible logger](http://www.php-fig.org/psr/psr-3/) like
-[Monolog](https://github.com/Seldaek/monolog).
-
-
-Installation
----
-
-The Jasny Error Handler package is available on [packagist](https://packagist.org/packages/jasny/error-handler).
-Install it using composer:
-
-    composer require jasny/error-handler-middleware
-
-
-Usage
----
-
-```php
-use Jasny\ErrorHandler;
-use Monolog\Logger;
-use Monolog\Handler\StreamHandler;
-
-$log = new Logger('test');
-$log->pushHandler(new StreamHandler('path/to/your.log'));
-
-$errorHandler = new ErrorHandler();
-
-// Log errors
-$errorHandler->setLogger($log);
-
-// Log fatal errors and warnings in addition to catchable errors
-$errorHandler->alsoLog(E_PARSE | E_ERROR | E_WARNING | E_USER_WARNING);
-
-// PHP 5 support
-$errorHandler->converErrorsToExceptions();
-```
-
 For example use it with **Relay**:
 
 ```php
@@ -60,8 +76,10 @@ use Relay\RelayBuilder;
 use Jasny\HttpMessage\ServerRequest;
 use Jasny\HttpMessage\Response;
 
+$errorHandler = new Jasny\ErrorHandler();
+
 $relay = new RelayBuilder();
-$dispatcher = $relay->newInstance([$errorHandler]);
+$dispatcher = $relay->newInstance([$errorHandler->asMiddleware()]);
 
 $response = $dispatcher((new ServerRequest())->withGlobalEnvironment(), new Response());
 ```
@@ -74,41 +92,33 @@ use Jasny\Router\Routes;
 use Jasny\HttpMessage\ServerRequest;
 use Jasny\HttpMessage\Response;
 
+$errorHandler = new Jasny\ErrorHandler();
+
 $routes = new Routes\Glob(['/**' => ['controller' => '$1', 'id' => '$2']);
 $router = new Router($routes);
 
-$router->add($errorHandler);
+$router->add($errorHandler->asMiddleware());
 
 $response = $dispatcher((new ServerRequest())->withGlobalEnvironment(), new Response());
 ```
 
-## Logging
+### PHP 5 support
 
-By default the error handler with only catch [Throwables](http://php.net/manual/en/class.throwable.php) and not set the
-[php error handler](http://php.net/set_error_handler).
-
-To log these errors, set the logger using the `setLogger()` method.
-
-The `alsoLog()` method will set the error handler, so warnings and notices can be logged. It may also [register a
-shutdown function](http://php.net/manual/en/function.register-shutdown-function.php) to handle uncatchable fatal
-errors.
-
-## PHP 5 support
-
-To add support for PHP5, you should call `converErrorsToExceptions()`. This method will convert an error to an
+With PHP 5 errors aren't thrown, so the middleware won't handle it. To add middleware support for errors in PHP5, you
+should call `converErrorsToExceptions()`. This method will convert an error to an
 [ErrorException](http://php.net/manual/en/class.errorexception.php).
 
 ## Handling fatal errors
 
-Some errors, like syntax errors, are not thrown and can't be handled be a user defined error handler. The PHP script
-will always exit right away when such an error occurs.
+Errors that are not thrown, like syntax errors, are not caught, will cause a fatal error. These errors will be logged
+by the error when specified with the `logUncaught()` method.
 
-By [registering a shutdown function](http://php.net/manual/en/function.register-shutdown-function.php) we can act upon
-these errors. Using `alsoLog(E_ERROR | E_PARSE)` will cause these errors to be logged. With the `onFatalError()` method
-you take additional action, like output an error message.
+With the `onFatalError()` method you take additional action, like output a pretty error message.
 
 ```php
 ob_start();
+
+$errorHandler = new Jasny\ErrorHandler();
 
 $errorHandler->onFatalError(function() {
     http_response_code(500);
